feat: ensure product quantity can be decreased

+++ src/actions/decrease-product-quantity/index.ts
@@ -7,15 +7,13 @@ import { db } from "@/db";
 import { cartItemTable } from "@/db/schema";
 import { auth } from "@/lib/auth";
 
-import {
-  RemoveProductFromCartSchema,
-  removeProductFromCartSchema,
-} from "./schema";
+import { decreaseCartProductQuantitySchema } from "./schema";
+import { DecreaseCartProductQuantitySchema } from "./schema";
 
-export const removeProductFromCart = async (
-  data: RemoveProductFromCartSchema
+export const decreaseProductFromCart = async (
+  data: DecreaseCartProductQuantitySchema
 ) => {
-  removeProductFromCartSchema.parse(data);
+  decreaseCartProductQuantitySchema.parse(data);
 
   const session = await auth.api.getSession({
     headers: await headers(),
@@ -33,12 +31,18 @@ export const removeProductFromCart = async (
     },
   });
 
-  if (cartItem?.cart.userId !== session.user.id) {
-    throw new Error("Unauthorized");
-  }
   if (!cartItem) {
     throw new Error("Product variant not found in cart");
   }
+  if (cartItem?.cart.userId !== session.user.id) {
+    throw new Error("Unauthorized");
+  }
+  if (cartItem.quantity === 1) {
+    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+  }
 
-  await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+  await db
+    .update(cartItemTable)
+    .set({ quantity: cartItem.quantity - 1 })
+    .where(eq(cartItemTable.id, cartItem.id));
 };
